Saving an idea for using ref() and watchEffect() outside of a Vue component
Playing with the idea of making a reactive ref that will hold an array of data that should be displayed on the cards in the scene. This ref would live in the scene file but could be set with data from Vue. The scene file would use a watchEffect to watch for changes to this ref, using any new data to update the cards in the scene.